% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_interface.R
\name{AphiaClassificationsByAphiaIDs}
\alias{AphiaClassificationsByAphiaIDs}
\title{Get taxonomy from aphia IDs}
\usage{
AphiaClassificationsByAphiaIDs(aphiaIDs)
}
\description{
Get taxonomy from aphia IDs
}
