% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_interface.R
\name{AphiaDistributionsByAphiaIDs}
\alias{AphiaDistributionsByAphiaIDs}
\title{Get Distribution from Aphia IDs}
\usage{
AphiaDistributionsByAphiaIDs(aphiaIDs)
}
\description{
Get Distribution from Aphia IDs
}
