% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_interface.R
\name{.safe_bind_rows}
\alias{.safe_bind_rows}
\title{Safe Bind Rows for API Schemas}
\usage{
.safe_bind_rows(df1, df2)
}
\description{
Some columns at certain API request will return empty values, such as retrieving direct children at family level, species will be empty. Merging the request result together where these columns are empty at one hand side and not empty on the other hand causes crashes with \code{bind_rows}. This function checks before hand and convert the empty3 (logical) column to type of the corresponding column in the other dara frame.
}
