% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_distribution}
\alias{get_distribution}
\title{Get distribution records of all Aphia IDs (species)}
\usage{
get_distribution(aphiaIDs)
}
\arguments{
\item{aphiaIDs}{A numeric vector of Aphia IDs.}
}
\value{
A dataframe of
}
\description{
Get distribution records of all Aphia IDs (species)
}
