% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get taxonomy (Records) from Aphia IDs}
\usage{
get_taxonomy(aphiaIDs)
}
\arguments{
\item{aphiaIDs}{A numeric vector of Aphia IDs.}
}
\value{
A dataframe with one row of taxonomic information.
}
\description{
Get taxonomy (Records) from Aphia IDs
}
