% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{request}
\alias{request}
\title{Execute Request}
\usage{
request(
  url,
  media = "application/json",
  query = list(),
  http_error = F,
  request_error = T,
  ...
)
}
\description{
Internal. Interface that execute communication with WoRMS webservice
}
