% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_interface.R
\name{AphiaChildrenByAphiaID}
\alias{AphiaChildrenByAphiaID}
\title{Retrieving Childrens from an Aphia ID}
\usage{
AphiaChildrenByAphiaID(
  aphiaID,
  recursive = T,
  rank = "Species",
  accept = T,
  marine = T,
  extant = T
)
}
\arguments{
\item{recursive}{logical. Whether to loop through children that are not species. If \code{TRUE} (default), all species belonging to given taxon is returned, if \code{FALSE}, only the direct children are returned.}

\item{rank}{Character. The lowest taxonomic rank to be retrieved. Default is "species".}

\item{accept}{logical. Whether to only return the accepted taxa.}

\item{marine}{logical. Whether to only return the marine taxa.}

\item{extant}{logical. Whether to only return the extant taxa.}
}
\description{
Retrieving Childrens from an Aphia ID
}
